Remove core/stop
Introducing it was a bad idea. Its meaning was unclear and there were no
guarantees that it would kill an actor which was waiting for a new message.

Hypothetically, actors could try to acquire their semaphores with a
timeout, check whether core/stop was called on them, and react
appropriately. Still, it wouldn't be a good idea because actors whose
callbacks take a longer while wouldn't be killed immediately. Such
abstraction is rather leaky.

What seems to make more sense is killing actors by defining a callback
which calls the die function, as in:

    (himilsbach.core/new
       msg    (do-something msg)
       [:die] (die))